+++ common/src/main/java/store/_0982/common/domain/settlement/SellerBalanceHistoryStatus.java
@@ -1,4 +1,4 @@
-package store._0982.common.domain.sellerbalance;
+package store._0982.common.domain.settlement;
 
 public enum SellerBalanceHistoryStatus {
     CREDIT,     // 입금/증가
